Move init plugins to pyplugins/init/ and make them user-extensible
The init plugin classes now live in pyplugins/init/ (18 files, grouped by
concern) and are discovered at init time rather than hardcoded:

- init_runner.discover_init_plugins() scans <plugin_path>/init/ (built-ins,
  zero Dockerfile changes since /pyplugins is already copied/mounted), the
  project's plugins.d/, and any --init-plugin-path directories. Later
  directories shadow earlier ones by class name, so a project can override a
  built-in. Broken plugin files are skipped with a warning. Files are imported
  through the existing introspection path, so `penguin schema <init_plugin>`
  works too.
- penguin init gains --jobs, --init-plugin-path, --enable, --disable,
  threaded through fakeroot_gen_config to the gen_config CLI. --enable
  force-enables disabled-by-default patches (e.g. root_shell); --disable
  skips a plugin entirely.
- static_analyses.py and config_patchers.py are reduced to shared helpers
  (FileSystemHelper, find_kernel_versions, TarHelper, FileHelper,
  NvramHelper, make_shims, RESOURCES) that the plugins import from core.
  ShimBinaries became the make_shims() helper function.
- SingleShot is now discovered like everything else (previously defined but
  never registered); it stays disabled by default so it only adds a disabled
  patch file. ShimFwEnv (unfinished, raised NotImplementedError on
  construction) is dropped; git history has it if it's ever finished.
- Output verified identical to the step-1 golden baseline on a fixture
  firmware, modulo the new disabled single_shot.yaml.
- New tests: discovery of all 38 built-ins, plugins.d extension and
  shadowing, --disable filtering, broken-file tolerance, and a pyplugins
  basename-uniqueness lint (the recursive plugin lookup rejects ambiguous
  names).
